anomaly_detector/requirements.txt
huggingface-hub
keras
numpy
pandas==1.5.3
python-dotenv
quixstreams~=2.8.0
scikit-learn
tensorflow